pyproject.toml
[tool.poetry]
name = "celebit-backend-django"
version = "0.1.0"
description = ""
authors = ["Einsicht1 <kpl5672@gmail.com>"]
readme = "README.md"
packages = [{include = "celebit_backend_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
mysqlclient = "^2.1.1"
gunicorn = "^20.1.0"
black = "^23.3.0"
pre-commit = "^3.3.2"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
